+++ CHANGELOG.md
@@ -16,8 +16,9 @@
 * JSONObject支持初始大小和按照KEY有序（感谢【深圳】Vmo）
 * 对ImageUtil重构以支持更对类型参数
 * DateTime增加offsetNew方法
-* 增加Range类，抽象递增递减列表
+* 增加Range类，抽象递增递减列表（感谢@【悉尼】C - ActFramework）
 * 扩充XmlUtil，提供更多重载方法（感谢@【北京】仓山有井名为空）
+* NumberUtil增加decimalFormatMoney方法（感谢@【北京】宁静 ）
 
 ### Bug修复
 * 修复Entity.parseEntity方法中获取表名逻辑的问题（感谢@【北京】游弋苍茫）